requirements.txt
fastapi>=0.112,<1
uvicorn[standard]>=0.30,<1
httpx>=0.27,<1
pyodbc>=5.1,<6
sqlglot>=25,<26
python-dotenv>=1.0,<2
pydantic>=2.8,<3
python-multipart>=0.0.9,<0.1
azure-identity>=1.17,<2
openai>=1.51,<2
